requirements.txt
# requirements.txt

# Django – version 5.2 (ou compatible)
Django>=5.2,<6.0

# Django REST Framework pour la création d'API
djangorestframework>=3.12

# Gestion des variables d'environnement
python-dotenv>=0.15

# Connecteur PostgreSQL (adapté pour le paramétrage django.db.backends.postgresql)
psycopg2-binary>=2.9


# django-cors-headers pour gérer les politiques CORS
django-cors-headers>=3.13


# JWT pour authentification
djangorestframework-simplejwt>=5.2

# Type annotations
django-stubs

# Gestion des dates
python-dateutil>=2.8.2
types-python-dateutil

# Serveur web
gunicorn>=20.1.0

# Interface d'administration Django
django-admin-interface

pytz
types-pytz

kafka-python
requests>=2.28.0
types-requests
pydantic

# image processing
Pillow>=9.0

# S3/CDN storage
django-storages[boto3]>=1.12
boto3>=1.24

celery
redis
